Added `ser/enums` module
